% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-nearest neighbors cross-validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame.}

\item{cl}{true class value of your training data.}

\item{k_nn}{integer representing the number of neighbors.}

\item{k_cv}{integer representing the number of folds.}
}
\value{
List containing elements: class, a vector of the predicted class for
for all observations; and cv_err, a numeric with the cross-validation
misclassification error.
}
\description{
This function predicts output class of an input data using k-nearest
neighbors cross-validation
}
\examples{
# import palmerpenguins data
# delete rows with NA data
penguins <- na.omit(mypracticepackage::my_penguins)

# create x and y matrices
train <- penguins[3:6]
cl <- penguins \%>\% dplyr::pull(species)
cv_err <- my_knn_cv(train, cl, 1, 5)$cv_error

}
\keyword{prediction}
