% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model fit function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula class object, similar to \code{lm()}.}

\item{data}{input data frame.}
}
\value{
Table with rows for each coefficient (including the intercept) and
columns for the Estimate, Std. Error, t value, and Pr(>|t|).
}
\description{
This function fits linear models, and can be used to perform linear
regression.
}
\examples{
formula <- mpg ~ hp + wt
data <- mtcars
my_lm(formula, data)

}
