% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random forest cross-validation function.}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{number of folds}
}
\value{
Numeric indicating average MSE across all k folds
}
\description{
This function predicts \code{body_mass_g} using covariates
\code{bill_length_mm}, \code{bill_depth_mm}, and \code{flipper_length_mm}.
}
\examples{
my_rf_cv(5)

}
